# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contacts & Services API",
        "description": "API with Google OAuth-secured endpoints for managing contacts and services.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://crud-api-project-y9c8.onrender.com"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Contact": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "email",
                    "phone"
                ]
            },
            "Service": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "provider": {
                        "type": "string"
                    }
                },
                "required": [
                    "title",
                    "description",
                    "price",
                    "category",
                    "duration",
                    "available",
                    "provider"
                ]
            }
        }
    },
    "paths": {
        "/contacts": {
            "get": {
                "summary": "Get all contacts",
                "responses": {
                    "200": {
                        "description": "List of contacts"
                    }
                }
            },
            "post": {
                "summary": "Create a new contact",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized: Please login to access this resource."
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "summary": "Get contact by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact found"
                    },
                    "404": {
                        "description": "Contact not found"
                    }
                }
            },
            "put": {
                "summary": "Update contact by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized: Please login to access this resource."
                    },
                    "404": {
                        "description": "Contact not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete contact by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact deleted"
                    },
                    "401": {
                        "description": "Unauthorized: Please login to access this resource."
                    },
                    "404": {
                        "description": "Contact not found"
                    }
                }
            }
        },
        "/services": {
            "get": {
                "summary": "Get all services",
                "responses": {
                    "200": {
                        "description": "List of services"
                    }
                }
            },
            "post": {
                "summary": "Create a new service",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Service"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized: Please login to access this resource."
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "summary": "Get service by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service found"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            },
            "put": {
                "summary": "Update service by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Service"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized: Please login to access this resource."
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete service by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted"
                    },
                    "401": {
                        "description": "Unauthorized: Please login to access this resource."
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            }
        }
    }
}